// submodules/ui-coverage-report/src/Models/Actions.ts
export enum ActionType {
  // input
  Fill = 'FILL',
  Type = 'TYPE',
  Select = 'SELECT',

  // action
  Click = 'CLICK',
  Hover = 'HOVER',

  // assert
  Text = 'TEXT',
  Value = 'VALUE',
  Hidden = 'HIDDEN',
  Visible = 'VISIBLE',
  Checked = 'CHECKED',
  Enabled = 'ENABLED',
  Disabled = 'DISABLED',
  Unchecked = 'UNCHECKED'
}